+++ CHANGELOG.md
@@ -91,6 +91,10 @@ Volumes:
 
 Miscellaneous:
 
+- Neuron Search: name queries can now be marked as exact, which prevents
+  matching against parts of other names. This is ignored for regular
+  expressions.
+
 - Projects can now be kept open if all last stack-viewers are closed. To not
   automatically close the open project with the last stack-viewer, uncheck the
   "Project closes with last stack viewer" option in the Settings Widget.